+++ README.md
@@ -45,18 +45,9 @@ The processing follows a structured organization:
 - **Build Directory:** A local mirror of the S3 storage, structured similarly for consistency.
 
 ```plaintext
-PREVIOUS STRUCTURE:
 # Example directory structure
 BUILD_DIR/BUCKET/NEWSPAPER/<NEWSPAPER-YEAR>.jsonl.bz2
 BUILD_DIR/BUCKET/PROCESSING_TYPE/VERSION/NEWSPAPER/<NEWSPAPER-YEAR>.jsonl.bz2
-```
-
-
-```plaintext
-NEW STRUCTURE:
-# Example directory structure
-BUILD_DIR/BUCKET/PREFIX/PROVIDER/NEWSPAPER/<NEWSPAPER-YEAR>.jsonl.bz2
-BUILD_DIR/BUCKET/PROCESSING_TYPE_VERSION/PROVIDER/NEWSPAPER/<NEWSPAPER-YEAR>.jsonl.bz2
 ```
 
 ## Setup
@@ -100,9 +91,6 @@ BUILD_DIR/BUCKET/PROCESSING_TYPE_VERSION/PROVIDER/NEWSPAPER/<NEWSPAPER-YEAR>.jso
    ```bash
    make setup
    ```
-   
-6. Make configuration changes as needed in `local.config.mk`, following the example `config.local.sample.mk`.
-
 
 ## Usage
