+++ config.local.sample.mk
@@ -23,54 +23,24 @@ $(info Make: Including config.local.mk: $(shell readlink -f config.local.mk))
 # You can set two of them here! OUT_S3_BUCKET_PROCESSED_DATA and OUT_S3_PROCESSED_VERSION
 # Don't change the OUT_S3_PROCESSED_INFIX!
 # : s3://$OUT_S3_BUCKET_PROCESSED_DATA/$OUT_S3_PROCESSED_INFIX/$OUT_S3_PROCESSED_VERSION
-#OUT_S3_BUCKET_PROCESSED_DATA := 40-processed-data-sandbox
-#OUT_S3_PROCESSED_VERSION := v1.0.1
-#
-## Were to write the local files
-#BUILD_DIR ?= build.d
-#
-## HUGGINGFACE MODEL SETTINGS
-## set the model cache directory to a local project directory (default:
-## ~/.cache/huggingface/transformers/)
-## should be an fast local disk
-#HF_HOME ?= ./hf.d
-#
-## set the number of parallel jobs when processing every newspaper (each newspaper-year
-## is on job; the default is 2)
-#MAKE_PARALLEL_OPTION ?= --jobs 2
-#
-## If you want to restrict the newspaper to work on
-#NEWSPAPER ?= SNL/EXP
-#
-## suppress the logging output of make itself
-#LOGGING_LEVEL := WARNING
+OUT_S3_BUCKET_PROCESSED_DATA := 40-processed-data-sandbox
+OUT_S3_PROCESSED_VERSION := v1.0.1
 
-# Inform
-$(info Make: Including config.local.mk: $(shell readlink -f config.local.mk))
-
-# Provider + Newspaper (pair)
-PROVIDER  ?= SNL
-NEWSPAPER ?= EXP
-
-# Set the model name and version
-CREATOR_NAME ?= Alibaba-NLP
-HF_MODEL_NAME ?= gte-multilingual-base
-HF_MODEL_VERSION ?= f7d567e
-HF_FULL_MODEL_NAME ?= $(CREATOR_NAME)/$(HF_MODEL_NAME)
-  $(call log.debug, HF_FULL_MODEL_NAME)
-
-# S3 rebuilt input bucket (MUST NOT contain '/' !!!)
-IN_S3_BUCKET_REBUILT := 000-processing-test-samples
-
-# Prefix UNDER the bucket
-IN_S3_PREFIX_REBUILT := lingproc/lingproc-test-v1.0.0
-
-# HF etc.
-OUT_S3_BUCKET_PROCESSED_DATA := 000-processing-test-samples
-OUT_S3_PROCESSED_INFIX := textembeddings-$(HF_MODEL_NAME)
-OUT_S3_PROCESSED_VERSION := v0.0.1
-
+# Were to write the local files
 BUILD_DIR ?= build.d
+
+# HUGGINGFACE MODEL SETTINGS
+# set the model cache directory to a local project directory (default:
+# ~/.cache/huggingface/transformers/)
+# should be an fast local disk
 HF_HOME ?= ./hf.d
+
+# set the number of parallel jobs when processing every newspaper (each newspaper-year
+# is on job; the default is 2)
 MAKE_PARALLEL_OPTION ?= --jobs 2
+
+# If you want to restrict the newspaper to work on
+NEWSPAPER ?= actionfem
+
+# suppress the logging output of make itself 
 LOGGING_LEVEL := WARNING
